bug fix
Single column return on mutual information when registration vector is poor attempted to combine integers of different classes. Fixed for same class.

+++ shared/adjust_MI.m
@@ -22,10 +22,10 @@ else
     %
     if strcmp(opt, 'rigid')
         coords = cast(coords, 'single') / scaling_factor + ...
-            search_region([3,1]);
+            double(search_region([3,1]));
     elseif strcmp(opt, 'initial_transformation')
         coords = double(coords) / scaling_factor + ...
-            search_region([3,1]);
+            double(search_region([3,1]));
     end
     %
 end
